container fixes & added get users api
