+++ .dockerignore
@@ -1,0 +1,4 @@
+./node_modules
+.dockerignore
+Dockerfile
+docker-compose.yaml
