+++ README.md
@@ -12,3 +12,12 @@ docker-compose -f docker-compose.yaml up
 
 chmod +x ./install.sh
 ./install.sh
+
+### run docker compose
+
+npm run docker
+
+### create a new .env file and paste the below two lines
+
+PORT=3000
+mongodb_uri=mongodb://localhost:27017/
