+++ public/constant.js
@@ -1,17 +1,17 @@
 /* eslint-disable-next-line */
 const APP = {
-    NAME: "test_app",
+    NAME: "bauth",
     VERSION: "1.0.0",
     BUILD: "0.0.1",
 };
 
 const DATABASE = {
-    NAME: "test_database",
+    NAME: "bauthdb",
     VERSION: "1.0.0",
 };
 
 const COLLECTION = {
-    ACCOUNTS: "test_accounts",
+    USERS: "users",
 };
 
 module.exports = { APP, DATABASE, COLLECTION };
