+++ server.js
@@ -2,12 +2,13 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const morgan = require("morgan");
 require("./logger").intialize();
+require("dotenv").config();
 const logger = require("./logger").logger;
 const path = require("path");
 const mongo_conn_native = require("./mongo_conn_native").Connection;
 
 // Routes
-const routeTestAPI = require("./restAPI/routes/route_test");
+const routeUserAPI = require("./restAPI/routes/route_user");
 
 // Initialzie Express
 const app = express();
@@ -30,31 +31,31 @@ app.use(bodyParser.json({ limit: "500mb", extended: true }));
 
 // For CORS
 app.use((req, res, next) => {
-	res.header("Access-Control-Allow-Origin", "*");
-	res.header("Access-Control-Allow-Methods", "*");
-	res.header("Access-Control-Allow-Headers", "*");
-	res.header("Access-Control-Allow-Credentials", "true");
-	next();
+    res.header("Access-Control-Allow-Origin", "*");
+    res.header("Access-Control-Allow-Methods", "*");
+    res.header("Access-Control-Allow-Headers", "*");
+    res.header("Access-Control-Allow-Credentials", "true");
+    next();
 });
 
 mongo_conn_native.connectToMongo().then(
-	async() => {
-		// Routes
+    async() => {
+        // Routes
 
-		// testing APIs
-		app.use("/api/v1/testAPI", routeTestAPI);
+        // testing APIs
+        app.use("/api/v1/users", routeUserAPI);
 
-		let port = process.env.PORT || 3016;
-		app.listen(port, async() => {
-			logger.info(`Test Node is listening on port ${port}`);
-			console.log(`Test Node is listening on port ${port}`);
-		});
-	},
-	(err) => {
-		console.log("Unable to connect mongo");
-		console.log(err);
-		logger.error(err);
-	}
+        let port = process.env.PORT || 3016;
+        app.listen(port, async() => {
+            logger.info(`Test Node is listening on port ${port}`);
+            console.log(`Test Node is listening on port ${port}`);
+        });
+    },
+    (err) => {
+        console.log("Unable to connect mongo");
+        console.log(err);
+        logger.error(err);
+    }
 );
 
 exports.app = app;
